ch11/echo.c: 修复一个编译警告

+++ sample/ch11/echo.c
@@ -16,7 +16,7 @@ void echo(int connfd)
 
         Rio_readinitb(&rio, connfd);
         while ((n = Rio_readlineb(&rio, buf, MAXLINE)) != 0) {
-                printf("server received %d bytes\n", n);
+                printf("server received %d bytes\n", (int)n);
                 Rio_writen(connfd, buf, n);
         }
 }
